+++ dev/com.ibm.ws.concurrent.mp.1.0/bnd.bnd
@@ -23,8 +23,7 @@ Require-Capability: osgi.ee;filter:="(&(osgi.ee=JavaSE)(version=1.8))"
 WS-TraceGroup: concurrent
 
 Private-Package:\
-  com.ibm.ws.concurrent.mp,\
-  com.ibm.ws.concurrent.rx
+  com.ibm.ws.concurrent.mp
     
 instrument.classesExcludes: com/ibm/ws/concurrent/mp/resources/*.class
