--- dev/com.ibm.ws.concurrent.mp.1.0/src/com/ibm/ws/concurrent/rx/package-info.java
@@ -1,18 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2017 IBM Corporation and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *     IBM Corporation - initial API and implementation
- *******************************************************************************/
-/**
- * @version 1.0
- */
-@org.osgi.annotation.versioning.Version("1.0")
-@TraceOptions(traceGroup = "concurrent")
-package com.ibm.ws.concurrent.rx;
-
-import com.ibm.websphere.ras.annotation.TraceOptions;
