+++ dev/com.ibm.ws.concurrent_fat_rx/publish/files/features/managedCompletableFuture-1.0.mf
@@ -4,4 +4,3 @@ Subsystem-Version: 1.0
 Subsystem-Content: com.ibm.ws.concurrent.mp.1.0
 Subsystem-Type: osgi.subsystem.feature
 IBM-Feature-Version: 2
-IBM-API-Package: com.ibm.ws.concurrent.rx; type="ibm-api"
